SWAR-34 Correct Logger
Actully tested this time unlike last time. Uses the logMessage method to log this message. Obtains all the data from the initialize method above. I used a next line command to make the output more readable on a smaller console and tried to format the data as neatly as possible.