SWAR-68
replaced the fast, slow and const PID functions with a universal one. Looking at the code, I do not believe that a PID class object is needed for this code, as the only time in DriveController that it is ever called is after the type of PID has been verified by a private result object.