SWAR-70
Cleaned up the code a little and rearranged a few things we changed so it reads better and is shorter.

Fixed obstacle avoidance to check if you are still trying to get home or if you are just trying to search. robot now navigates obstacles if trying to get home.

Fixed pickup so robot now goes straight into waypoint home navigation instead of spinng around.

No longer stop working after picking up a cube.

Robot will still spin a bit more when exiting home after drop off. This was observed constistantly but the effects it has seem to be more beneficial. It found nearby cubes (it spun 360 degrees sometimes) and put cubes back in that it knocked out.