SWAR-35 Select target based on block distance from center of camera
This commit modifies PickUpController::SetTagData to select the target block for pickup based on its absolute positionX as opposed to its distance from the camera. The idea is to reduce the amount of angular error from the bot to the target to make pickup more efficient and reduce thrashing between targets in clusters.